Chewy::Index::Actions specs are split per method
