+++ spec/chewy/index/actions/delete_bang_spec.rb
@@ -1,0 +1,72 @@
+require 'spec_helper'
+
+describe Chewy::Index::Actions do
+  before { drop_indices }
+
+  before do
+    stub_index :dummies
+    stub_index :dummies_suffixed
+  end
+
+  describe '.delete!' do
+    specify { expect { DummiesIndex.delete! }.to raise_error(Elastic::Transport::Transport::Errors::NotFound) }
+    specify do
+      expect do
+        DummiesIndex.delete!('2013')
+      end.to raise_error(Elastic::Transport::Transport::Errors::NotFound)
+    end
+
+    context do
+      before do
+        DummiesIndex.create
+        DummiesSuffixedIndex.create 'should_not_appear'
+      end
+
+      specify { expect(DummiesIndex.delete!['acknowledged']).to eq(true) }
+
+      context do
+        before { DummiesIndex.delete! }
+        specify { expect(Chewy.client.indices.exists(index: 'dummies')).to eq(false) }
+        specify { expect(Chewy.client.indices.exists(index: 'dummies_suffixed')).to eq(true) }
+      end
+    end
+
+    context do
+      before do
+        DummiesIndex.create '2013'
+        DummiesSuffixedIndex.create 'should_not_appear'
+      end
+
+      specify { expect(DummiesIndex.delete!('2013')['acknowledged']).to eq(true) }
+
+      context do
+        before { DummiesIndex.delete!('2013') }
+        specify { expect(Chewy.client.indices.exists(index: 'dummies')).to eq(false) }
+        specify { expect(Chewy.client.indices.exists(index: 'dummies_2013')).to eq(false) }
+        specify { expect(Chewy.client.indices.exists(index: 'dummies_suffixed')).to eq(true) }
+      end
+
+      context do
+        before { DummiesIndex.create '2014' }
+
+        specify { expect(DummiesIndex.delete!['acknowledged']).to eq(true) }
+
+        context do
+          before { DummiesIndex.delete! }
+          specify { expect(Chewy.client.indices.exists(index: 'dummies')).to eq(false) }
+          specify { expect(Chewy.client.indices.exists(index: 'dummies_2013')).to eq(false) }
+          specify { expect(Chewy.client.indices.exists(index: 'dummies_2014')).to eq(false) }
+          specify { expect(Chewy.client.indices.exists(index: 'dummies_suffixed')).to eq(true) }
+        end
+
+        context do
+          before { DummiesIndex.delete!('2014') }
+          specify { expect(Chewy.client.indices.exists(index: 'dummies')).to eq(true) }
+          specify { expect(Chewy.client.indices.exists(index: 'dummies_2013')).to eq(true) }
+          specify { expect(Chewy.client.indices.exists(index: 'dummies_2014')).to eq(false) }
+          specify { expect(Chewy.client.indices.exists(index: 'dummies_suffixed')).to eq(true) }
+        end
+      end
+    end
+  end
+end
