+++ spec/chewy/index/actions/purge_bang_spec.rb
@@ -1,0 +1,57 @@
+require 'spec_helper'
+
+describe Chewy::Index::Actions do
+  before { drop_indices }
+
+  before do
+    stub_index :dummies
+    stub_index :dummies_suffixed
+  end
+
+  describe '.purge!' do
+    specify { expect(DummiesIndex.purge!['acknowledged']).to eq(true) }
+    specify { expect(DummiesIndex.purge!('2013')['acknowledged']).to eq(true) }
+
+    context do
+      before { DummiesIndex.purge! }
+      specify { expect(DummiesIndex).to be_exists }
+      specify { expect(DummiesIndex.aliases).to eq([]) }
+      specify { expect(DummiesIndex.indexes).to eq(['dummies']) }
+
+      context do
+        before { DummiesIndex.purge! }
+        specify { expect(DummiesIndex).to be_exists }
+        specify { expect(DummiesIndex.aliases).to eq([]) }
+        specify { expect(DummiesIndex.indexes).to eq(['dummies']) }
+      end
+
+      context do
+        before { DummiesIndex.purge!('2013') }
+        specify { expect(DummiesIndex).to be_exists }
+        specify { expect(DummiesIndex.aliases).to eq(['dummies']) }
+        specify { expect(DummiesIndex.indexes).to eq(['dummies_2013']) }
+      end
+    end
+
+    context do
+      before { DummiesIndex.purge!('2013') }
+      specify { expect(DummiesIndex).to be_exists }
+      specify { expect(DummiesIndex.aliases).to eq(['dummies']) }
+      specify { expect(DummiesIndex.indexes).to eq(['dummies_2013']) }
+
+      context do
+        before { DummiesIndex.purge! }
+        specify { expect(DummiesIndex).to be_exists }
+        specify { expect(DummiesIndex.aliases).to eq([]) }
+        specify { expect(DummiesIndex.indexes).to eq(['dummies']) }
+      end
+
+      context do
+        before { DummiesIndex.purge!('2014') }
+        specify { expect(DummiesIndex).to be_exists }
+        specify { expect(DummiesIndex.aliases).to eq(['dummies']) }
+        specify { expect(DummiesIndex.indexes).to eq(['dummies_2014']) }
+      end
+    end
+  end
+end
